+++ imblearn/under_sampling/prototype_generation/tests/test_cluster_centroids.py
@@ -27,10 +27,13 @@ def test_fit_sample_auto():
     ratio = 'auto'
     cc = ClusterCentroids(ratio=ratio, random_state=RND_SEED)
     X_resampled, y_resampled = cc.fit_sample(X, Y)
-    X_gt = np.array([[0.92923648, 0.76103773], [0.47104475, 0.44386323],
-                     [0.13347175, 0.12167502], [0.06738818, -0.529627],
-                     [0.17901516, 0.69860992], [0.094035, -2.55298982]])
-    y_gt = np.array([0, 0, 0, 1, 1, 1])
+    X_gt = np.array([[0.06738818, -0.529627],
+                     [0.17901516, 0.69860992],
+                     [0.094035, -2.55298982],
+                     [0.92923648, 0.76103773],
+                     [0.47104475, 0.44386323],
+                     [0.13347175, 0.12167502]])
+    y_gt = np.array([1, 1, 1, 0, 0, 0])
     assert_allclose(X_resampled, X_gt, rtol=R_TOL)
     assert_array_equal(y_resampled, y_gt)
 
@@ -39,12 +42,16 @@ def test_fit_sample_half():
     ratio = .5
     cc = ClusterCentroids(ratio=ratio, random_state=RND_SEED)
     X_resampled, y_resampled = cc.fit_sample(X, Y)
-    X_gt = np.array([[0.92923648, 0.76103773], [0.47104475, 0.44386323],
-                     [0.13347175, 0.12167502], [0.09125309, -0.85409574],
-                     [0.19220316, 0.32337101], [0.094035, -2.55298982],
-                     [0.20792588, 1.49407907], [0.04352327, -0.20515826],
-                     [0.12372842, 0.6536186]])
-    y_gt = np.array([0, 0, 0, 1, 1, 1, 1, 1, 1])
+    X_gt = np.array([[0.09125309, -0.85409574],
+                     [0.19220316, 0.32337101],
+                     [0.094035, -2.55298982],
+                     [0.20792588, 1.49407907],
+                     [0.04352327, -0.20515826],
+                     [0.12372842, 0.6536186],
+                     [0.92923648, 0.76103773],
+                     [0.47104475, 0.44386323],
+                     [0.13347175, 0.12167502]])
+    y_gt = np.array([1, 1, 1, 1, 1, 1, 0, 0, 0])
     assert_allclose(X_resampled, X_gt, rtol=R_TOL)
     assert_array_equal(y_resampled, y_gt)
 
@@ -68,10 +75,13 @@ def test_fit_sample_object():
         ratio=ratio, random_state=RND_SEED, estimator=cluster)
 
     X_resampled, y_resampled = cc.fit_sample(X, Y)
-    X_gt = np.array([[0.92923648, 0.76103773], [0.47104475, 0.44386323],
-                     [0.13347175, 0.12167502], [0.06738818, -0.529627],
-                     [0.17901516, 0.69860992], [0.094035, -2.55298982]])
-    y_gt = np.array([0, 0, 0, 1, 1, 1])
+    X_gt = np.array([[0.06738818, -0.529627],
+                     [0.17901516, 0.69860992],
+                     [0.094035, -2.55298982],
+                     [0.92923648, 0.76103773],
+                     [0.47104475, 0.44386323],
+                     [0.13347175, 0.12167502]])
+    y_gt = np.array([1, 1, 1, 0, 0, 0])
     assert_allclose(X_resampled, X_gt, rtol=R_TOL)
     assert_array_equal(y_resampled, y_gt)