+++ imblearn/under_sampling/prototype_selection/nearmiss.py
@@ -11,6 +11,8 @@ from collections import Counter
 
 import numpy as np
 
+from sklearn.utils import safe_indexing
+
 from ..base import BaseUnderSampler
 from ...utils import check_neighbors_object
 from ...utils.deprecation import deprecate_parameter
@@ -154,10 +156,10 @@ NearMiss # doctest: +NORMALIZE_WHITESPACE
 
         Parameters
         ----------
-        X : ndarray, shape (n_samples, n_features)
+        X : {array-like, sparse matrix}, shape (n_samples, n_features)
             Original samples.
 
-        y : ndarray, shape (n_samples, )
+        y : array-like, shape (n_samples,)
             Associated label to X.
 
         dist_vec : ndarray, shape (n_samples, )
@@ -174,13 +176,7 @@ NearMiss # doctest: +NORMALIZE_WHITESPACE
 
         Returns
         -------
-        X_sel : ndarray, shape (num_samples, n_features)
-            Selected samples.
-
-        y_sel : ndarray, shape (num_samples, )
-            The associated label.
-
-        idx_sel : ndarray, shape (num_samples, )
+        idx_sel : ndarray, shape (num_samples,)
             The list of the indices of the selected samples.
 
         """
@@ -188,7 +184,9 @@ NearMiss # doctest: +NORMALIZE_WHITESPACE
         # Compute the distance considering the farthest neighbour
         dist_avg_vec = np.sum(dist_vec[:, -self.nn_.n_neighbors:], axis=1)
 
-        if dist_vec.shape[0] != X[y == key].shape[0]:
+        target_class_indices = np.flatnonzero(y == key)
+        if (dist_vec.shape[0] != safe_indexing(X,
+                                               target_class_indices).shape[0]):
             raise RuntimeError('The samples to be selected do not correspond'
                                ' to the distance matrix given. Ensure that'
                                ' both `X[y == key]` and `dist_vec` are'
@@ -243,18 +241,19 @@ NearMiss # doctest: +NORMALIZE_WHITESPACE
 
         Parameters
         ----------
-        X : ndarray, shape (n_samples, n_features)
+        X : {array-like, sparse matrix}, shape (n_samples, n_features)
             Matrix containing the data which have to be sampled.
 
-        y : ndarray, shape (n_samples, )
+        y : array-like, shape (n_samples,)
             Corresponding label for each sample in X.
 
         Returns
         -------
-        X_resampled : ndarray, shape (n_samples_new, n_features)
+        X_resampled : {ndarray, sparse matrix}, shape \
+(n_samples_new, n_features)
             The array containing the resampled data.
 
-        y_resampled : ndarray, shape (n_samples_new)
+        y_resampled : ndarray, shape (n_samples_new,)
             The corresponding label of `X_resampled`
 
         idx_under : ndarray, shape (n_samples, )
@@ -264,21 +263,20 @@ NearMiss # doctest: +NORMALIZE_WHITESPACE
         """
         self._validate_estimator()
 
-        X_resampled = np.empty((0, X.shape[1]), dtype=X.dtype)
-        y_resampled = np.empty((0, ), dtype=y.dtype)
-        if self.return_indices:
-            idx_under = np.empty((0, ), dtype=int)
+        idx_under = np.empty((0, ), dtype=int)
 
         target_stats = Counter(y)
         class_minority = min(target_stats, key=target_stats.get)
+        minority_class_indices = np.flatnonzero(y == class_minority)
 
-        self.nn_.fit(X[y == class_minority])
+        self.nn_.fit(safe_indexing(X, minority_class_indices))
 
         for target_class in np.unique(y):
             if target_class in self.ratio_.keys():
                 n_samples = self.ratio_[target_class]
-                X_class = X[y == target_class]
-                y_class = y[y == target_class]
+                target_class_indices = np.flatnonzero(y == target_class)
+                X_class = safe_indexing(X, target_class_indices)
+                y_class = safe_indexing(y, target_class_indices)
 
                 if self.version == 1:
                     dist_vec, idx_vec = self.nn_.kneighbors(
@@ -295,10 +293,10 @@ NearMiss # doctest: +NORMALIZE_WHITESPACE
                 elif self.version == 3:
                     self.nn_ver3_.fit(X_class)
                     dist_vec, idx_vec = self.nn_ver3_.kneighbors(
-                        X[y == class_minority])
+                        safe_indexing(X, minority_class_indices))
                     idx_vec_farthest = np.unique(idx_vec.reshape(-1))
-                    X_class_selected = X_class[idx_vec_farthest, :]
-                    y_class_selected = y_class[idx_vec_farthest]
+                    X_class_selected = safe_indexing(X_class, idx_vec_farthest)
+                    y_class_selected = safe_indexing(y_class, idx_vec_farthest)
 
                     dist_vec, idx_vec = self.nn_.kneighbors(
                         X_class_selected, n_neighbors=self.nn_.n_neighbors)
@@ -311,18 +309,12 @@ NearMiss # doctest: +NORMALIZE_WHITESPACE
             else:
                 index_target_class = slice(None)
 
-            X_resampled = np.concatenate(
-                (X_resampled, X[y == target_class][index_target_class]),
-                axis=0)
-            y_resampled = np.concatenate(
-                (y_resampled, y[y == target_class][index_target_class]),
-                axis=0)
-            if self.return_indices:
-                idx_under = np.concatenate(
-                    (idx_under, np.flatnonzero(y == target_class)[
-                        index_target_class]), axis=0)
+            idx_under = np.concatenate(
+                (idx_under, np.flatnonzero(y == target_class)[
+                    index_target_class]), axis=0)
 
         if self.return_indices:
-            return X_resampled, y_resampled, idx_under
+            return (safe_indexing(X, idx_under), safe_indexing(y, idx_under),
+                    idx_under)
         else:
-            return X_resampled, y_resampled
+            return safe_indexing(X, idx_under), safe_indexing(y, idx_under)